requirements.txt
aws-cdk-lib==2.81.0
constructs==10.2.35